chore: move eth_getLogs events fetching from explorer to bzx-common package
