+++ packages/bzx-common/src/domain/events/CloseWithDepositEvent.ts
@@ -1,29 +1,27 @@
 import { BigNumber } from '@0x/utils'
 import Asset from 'bzx-common/src/assets/Asset'
-import { timeStamp } from 'console'
 
-export class LiquidationEvent {
-  public static readonly topic0: string =
-    '0x46fa03303782eb2f686515f6c0100f9a62dabe587b0d3f5a4fc0c822d6e532d3'
+export default class CloseWithDepositEvent {
+  public static topic0: string =
+    '0x6349c1a02ec126f7f4fc6e6837e1859006e90e9901635c442d29271e77b96fb6'
   public readonly user: string //indexed
-  public readonly liquidator: string //indexed
+  public readonly lender: string //indexed
   public readonly loanId: string //indexed
-  public readonly lender: string
+  public readonly closer: string
   public readonly loanToken: Asset
   public readonly collateralToken: Asset
   public readonly repayAmount: BigNumber
   public readonly collateralWithdrawAmount: BigNumber
-  public readonly collateralToLoanRate: BigNumber // one unit of baseToken, denominated in quoteToken
+  public readonly collateralToLoanRate: BigNumber // one unit of loanToken, denominated in collateralToken
   public readonly currentMargin: BigNumber
   public readonly blockNumber: BigNumber
   public readonly txHash: string
-  public readonly timeStamp?: Date
 
   constructor(
     user: string,
-    liquidator: string,
-    loanId: string,
     lender: string,
+    loanId: string,
+    closer: string,
     loanToken: Asset,
     collateralToken: Asset,
     repayAmount: BigNumber,
@@ -31,21 +29,19 @@ export class LiquidationEvent {
     collateralToLoanRate: BigNumber,
     currentMargin: BigNumber,
     blockNumber: BigNumber,
-    txHash: string,
-    timeStamp?: Date
+    txHash: string
   ) {
     this.user = user
-    this.liquidator = liquidator
-    this.loanId = loanId
-    this.lender = lender
     this.loanToken = loanToken
     this.collateralToken = collateralToken
+    this.lender = lender
+    this.closer = closer
+    this.loanId = loanId
     this.repayAmount = repayAmount
     this.collateralWithdrawAmount = collateralWithdrawAmount
     this.collateralToLoanRate = collateralToLoanRate
     this.currentMargin = currentMargin
     this.blockNumber = blockNumber
     this.txHash = txHash
-    this.timeStamp = timeStamp
   }
 }
