+++ packages/bzx-common/src/domain/events/CloseWithSwapEvent.ts
@@ -1,7 +1,7 @@
 import { BigNumber } from '@0x/utils'
 import Asset from 'bzx-common/src/assets/Asset'
 
-export class CloseWithSwapEvent {
+export default class CloseWithSwapEvent {
   public static topic0: string =
     '0x2ed7b29b4ca95cf3bb9a44f703872a66e6aa5e8f07b675fa9a5c124a1e5d7352'
   public readonly user: string //indexed
