+++ packages/bzx-common/src/domain/events/TradeEvent.ts
@@ -1,7 +1,7 @@
 import { BigNumber } from '@0x/utils'
 import Asset from 'bzx-common/src/assets/Asset'
 
-export class TradeEvent {
+export default class TradeEvent {
   public static readonly topic0: string =
     '0xf640c1cfe1a912a0b0152b5a542e5c2403142eed75b06cde526cee54b1580e5c'
   public readonly user: string //indexed
