+++ packages/bzx-common/src/domain/events/index.ts
@@ -1,0 +1,19 @@
+import BorrowEvent from './BorrowEvent'
+import BurnEvent from './BurnEvent'
+import CloseWithDepositEvent from './CloseWithDepositEvent'
+import CloseWithSwapEvent from './CloseWithSwapEvent'
+import LiquidationEvent from './LiquidationEvent'
+import MintEvent from './MintEvent'
+import RolloverEvent from './RolloverEvent'
+import TradeEvent from './TradeEvent'
+
+export {
+  BorrowEvent,
+  BurnEvent,
+  CloseWithDepositEvent,
+  CloseWithSwapEvent,
+  LiquidationEvent,
+  MintEvent,
+  RolloverEvent,
+  TradeEvent
+}
