+++ packages/protocol-explorer/src/pages/LiquidationsPage.tsx
@@ -9,7 +9,7 @@ import { TxGrid } from '../components/TxGrid'
 import { ITxRowProps } from '../components/TxRow'
 import { UnhealthyChart } from '../components/UnhealthyChart'
 import Asset from 'bzx-common/src/assets/Asset'
-import { LiquidationEvent } from '../domain/LiquidationEvent'
+import { LiquidationEvent } from 'bzx-common/src/domain/events'
 import { LiquidationRequest } from '../domain/LiquidationRequest'
 import { Header } from '../layout/Header'
 import { ExplorerProviderEvents } from '../services/events/ExplorerProviderEvents'
